refactor(frontend): split FilterButtons into smaller components
