+++ apps/web/components/breadcrumb-filter/FilterButtons.tsx
@@ -1,10 +1,12 @@
 'use client';
 
-import React, { useState, useCallback } from 'react';
-import { Filter, Truck, Star, DollarSign, HardDrive, Cpu, Monitor, Zap, Palette, Settings, Brain } from 'lucide-react';
+import React, { useState } from 'react';
 import { cn } from '@/lib/utils';
 import { BreadcrumbFilter } from './BreadcrumbFilter';
-import { useRouter } from 'next/navigation';
+import { FilterButton } from './FilterButton';
+import { evFilterButtons } from './EvFilters';
+import { batteryFilterButtons } from './BatteryFilters';
+import { useBreadcrumb } from './useBreadcrumb';
 
 interface BreadcrumbItem {
   label: string;
@@ -21,44 +23,6 @@ interface FilterButtonsProps {
   onSubcategoryChange?: (setSubcategory: (subcategory: string) => void) => void;
 }
 
-interface FilterButtonProps {
-  label: string;
-  icon?: React.ReactNode;
-  isActive?: boolean;
-  hasDropdown?: boolean;
-  hasInfo?: boolean;
-  onClick?: () => void;
-}
-
-function FilterButton({ label, icon, isActive, hasDropdown, hasInfo, onClick }: FilterButtonProps) {
-  return (
-    <button
-      onClick={onClick}
-      className={cn(
-        'flex items-center gap-2 px-4 py-2 rounded-lg border transition-all duration-200 text-sm font-medium',
-        isActive
-          ? 'border-red-500 text-red-500 bg-red-50' // Nút "Bộ lọc" - màu đỏ
-          : 'border-gray-300 text-gray-700 bg-gray-50 hover:bg-gray-100 hover:border-gray-400'
-      )}
-    >
-      {icon && <span className="flex-shrink-0">{icon}</span>}
-      <span>{label}</span>
-      {hasDropdown && (
-        <svg className="h-3 w-3 text-gray-500" fill="none" viewBox="0 0 24 24" stroke="currentColor">
-          <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M19 9l-7 7-7-7" />
-        </svg>
-      )}
-      {hasInfo && (
-        <svg className="h-3 w-3 text-gray-500" fill="none" viewBox="0 0 24 24" stroke="currentColor">
-          <circle cx="12" cy="12" r="10" strokeWidth={2} />
-          <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M12 16v-4" />
-          <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M12 8h.01" />
-        </svg>
-      )}
-    </button>
-  );
-}
-
 export function FilterButtons({
   className,
   breadcrumbItems = [],
@@ -68,61 +32,9 @@ export function FilterButtons({
   onSubcategoryChange
 }: FilterButtonsProps) {
   const [activeFilter, setActiveFilter] = useState('all');
-  const router = useRouter();
 
-  // Breadcrumb state
-  const [breadcrumbState, setBreadcrumbState] = useState<{
-    items: BreadcrumbItem[];
-    currentCategory: string | null;
-    currentSubcategory: string | null;
-  }>({
-    items: [],
-    currentCategory: null,
-    currentSubcategory: null,
-  });
-
-  // Breadcrumb functions
-  const setCategory = useCallback((category: string) => {
-    setBreadcrumbState(prev => ({
-      ...prev,
-      currentCategory: category,
-      currentSubcategory: null,
-      items: [
-        {
-          label: category,
-          onClick: () => {
-            if (type === 'ev' && category === 'Xe điện') {
-              router.push('/posts/ev');
-            } else if (type === 'battery' && category === 'Pin EV') {
-              router.push('/posts/battery');
-            }
-          },
-        }
-      ]
-    }));
-  }, [type, router]);
-
-  const setSubcategory = useCallback((subcategory: string) => {
-    setBreadcrumbState(prev => ({
-      ...prev,
-      currentSubcategory: subcategory,
-      items: [
-        {
-          label: prev.currentCategory || 'Trang chủ',
-          onClick: () => {
-            if (type === 'ev' && prev.currentCategory === 'Xe điện') {
-              router.push('/posts/ev');
-            } else if (type === 'battery' && prev.currentCategory === 'Pin EV') {
-              router.push('/posts/battery');
-            }
-          },
-        },
-        {
-          label: subcategory,
-        }
-      ]
-    }));
-  }, [type, router]);
+  // Use breadcrumb hook
+  const { breadcrumbState, setCategory, setSubcategory } = useBreadcrumb(type);
 
   // Initialize breadcrumb on mount
   React.useEffect(() => {
@@ -143,110 +55,36 @@ export function FilterButtons({
   // Use internal breadcrumb if no external breadcrumbItems provided
   const displayBreadcrumbItems = breadcrumbItems.length > 0 ? breadcrumbItems : breadcrumbState.items;
 
-  // Filter buttons cho EV
-  const evFilterButtons = [
-    {
-      label: 'Bộ lọc',
-      icon: <Filter className="h-4 w-4" />,
-      isActive: activeFilter === 'all',
-      onClick: () => setActiveFilter('all')
-    },
-    {
-      label: 'Sẵn hàng',
-      icon: <Truck className="h-4 w-4" />,
-      isActive: activeFilter === 'available',
-      onClick: () => setActiveFilter('available')
-    },
-    {
-      label: 'Hàng mới về',
-      icon: <Star className="h-4 w-4" />,
-      isActive: activeFilter === 'new-arrivals',
-      onClick: () => setActiveFilter('new-arrivals')
-    },
-    {
-      label: 'Xem theo giá',
-      icon: <DollarSign className="h-4 w-4" />,
-      hasDropdown: true,
-      hasInfo: true,
-      isActive: activeFilter === 'price',
-      onClick: () => setActiveFilter('price')
-    },
-    {
-      label: 'Nhu cầu sử dụng',
-      hasDropdown: true,
-      hasInfo: true,
-      isActive: activeFilter === 'usage',
-      onClick: () => setActiveFilter('usage')
-    },
-    {
-      label: 'Quãng đường di chuyển',
-      hasDropdown: true,
-      hasInfo: true,
-      isActive: activeFilter === 'range',
-      onClick: () => setActiveFilter('range')
-    }
-  ];
-
-  // Filter buttons cho Battery
-  const batteryFilterButtons = [
-    {
-      label: 'Bộ lọc',
-      icon: <Filter className="h-4 w-4" />,
-      isActive: activeFilter === 'all',
-      onClick: () => setActiveFilter('all')
-    },
-    {
-      label: 'Sẵn hàng',
-      icon: <Truck className="h-4 w-4" />,
-      isActive: activeFilter === 'available',
-      onClick: () => setActiveFilter('available')
-    },
-    {
-      label: 'Hàng mới về',
-      icon: <Star className="h-4 w-4" />,
-      isActive: activeFilter === 'new-arrivals',
-      onClick: () => setActiveFilter('new-arrivals')
-    },
-    {
-      label: 'Xem theo giá',
-      icon: <DollarSign className="h-4 w-4" />,
-      hasDropdown: true,
-      hasInfo: true,
-      isActive: activeFilter === 'price',
-      onClick: () => setActiveFilter('price')
-    },
-    {
-      label: 'Dung lượng pin',
-      icon: <Zap className="h-4 w-4" />,
-      hasDropdown: true,
-      hasInfo: true,
-      isActive: activeFilter === 'capacity',
-      onClick: () => setActiveFilter('capacity')
-    },
-    {
-      label: 'Tình trạng pin',
-      hasDropdown: true,
-      hasInfo: true,
-      isActive: activeFilter === 'health',
-      onClick: () => setActiveFilter('health')
-    },
-    {
-      label: 'Số chu kỳ',
-      hasDropdown: true,
-      hasInfo: true,
-      isActive: activeFilter === 'cycles',
-      onClick: () => setActiveFilter('cycles')
-    },
-    {
-      label: 'Thương hiệu',
-      hasDropdown: true,
-      hasInfo: true,
-      isActive: activeFilter === 'brand',
-      onClick: () => setActiveFilter('brand')
-    }
-  ];
-
+  // Get filter buttons based on type
   const filterButtons = type === 'battery' ? batteryFilterButtons : evFilterButtons;
+
+  // Update button states with active filter
+  const buttonsWithState = filterButtons.map(button => ({
+    ...button,
+    isActive: button.label === 'Bộ lọc' ? activeFilter === 'all' :
+              button.label === 'Sẵn hàng' ? activeFilter === 'available' :
+              button.label === 'Hàng mới về' ? activeFilter === 'new-arrivals' :
+              button.label === 'Xem theo giá' ? activeFilter === 'price' :
+              button.label === 'Nhu cầu sử dụng' ? activeFilter === 'usage' :
+              button.label === 'Quãng đường di chuyển' ? activeFilter === 'range' :
+              button.label === 'Dung lượng pin' ? activeFilter === 'capacity' :
+              button.label === 'Tình trạng pin' ? activeFilter === 'health' :
+              button.label === 'Số chu kỳ' ? activeFilter === 'cycles' :
+              button.label === 'Thương hiệu' ? activeFilter === 'brand' : false,
+    onClick: () => {
+      const filterKey = button.label === 'Bộ lọc' ? 'all' :
+                       button.label === 'Sẵn hàng' ? 'available' :
+                       button.label === 'Hàng mới về' ? 'new-arrivals' :
+                       button.label === 'Xem theo giá' ? 'price' :
+                       button.label === 'Nhu cầu sử dụng' ? 'usage' :
+                       button.label === 'Quãng đường di chuyển' ? 'range' :
+                       button.label === 'Dung lượng pin' ? 'capacity' :
+                       button.label === 'Tình trạng pin' ? 'health' :
+                       button.label === 'Số chu kỳ' ? 'cycles' :
+                       button.label === 'Thương hiệu' ? 'brand' : 'all';
+      setActiveFilter(filterKey);
+    }
+  }));
 
   return (
     <div className={cn('w-full bg-white border-b border-gray-200', className)}>
@@ -259,7 +97,7 @@ export function FilterButtons({
         <div className="space-y-4">
           <h3 className="text-lg font-bold text-gray-900">Chọn theo tiêu chí</h3>
           <div className="flex flex-wrap gap-3">
-            {filterButtons.map((button, index) => (
+            {buttonsWithState.map((button, index) => (
               <FilterButton
                 key={index}
                 label={button.label}
