+++ apps/web/components/breadcrumb-filter/index.ts
@@ -1,2 +1,6 @@
 export { BreadcrumbFilter } from './BreadcrumbFilter';
 export { FilterButtons } from './FilterButtons';
+export { FilterButton } from './FilterButton';
+export { evFilterButtons } from './EvFilters';
+export { batteryFilterButtons } from './BatteryFilters';
+export { useBreadcrumb } from './useBreadcrumb';
